feat(router): renames 'dbInit' to 'init'
