+++ src/routes/index.ts
@@ -1,16 +1,14 @@
 import { Router } from 'express';
 import { home } from './home';
-import { greet } from './greet';
-import { dbInit } from './dbInit';
+import { dbInit } from './init';
 import { search } from './search';
-import { filter } from './filter';
+import { target } from './target';
 
 const router: Router = Router();
 
 router.get('/', home);
 router.get('/search', search);
-router.get('/filter', filter);
-router.post('/greet', greet);
+router.get('/target', target);
 router.post('/init', dbInit);
 
 export { router };
