+++ src/routes/init/index.ts
